+++ packages/args/src/helpers/isFlagGroup.ts
@@ -1,0 +1,5 @@
+const FLAG_GROUP = /^-[a-z]{2,}$/iu;
+
+export default function isFlagGroup(arg: string): boolean {
+  return FLAG_GROUP.test(arg);
+}
